refactor(viewer): move FEA session access, overlay visibility and teardown out of the loader
First cut of the fea/streaming package. The session accessor, the
plugin-facing selection adapters, the beam/edge/result-colour
visibility rule with the undeformed ghost, and clearActiveFeaStreaming
each get a module with a header stating what it owns and reads.

The loader's one remaining `let` -- whether result colours are on
screen, as distinct from the store toggle -- moves onto the session
handle as `resultColorsShown`: it is a fact about the loaded model and
goes with it when the session closes. Every export is re-exported from
load_fea_streaming.ts, so no importer changes.
